exec: template out colvec
This commit extracts the execgen template string from `colvec_gen.go` into its
own template file at `colvec_tmpl.go`. This makes it easier to make templating
changes.

Release note: None
