+++ pkg/sql/exec/execgen/cmd/execgen/colvec_gen.go
@@ -16,123 +16,34 @@ package main
 
 import (
 	"io"
+	"io/ioutil"
+	"strings"
 	"text/template"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/exec/types"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 )
 
-const colVecTemplate = `
-package exec
-
-import (
-  "fmt"
-	
-	"github.com/cockroachdb/apd"
-	"github.com/cockroachdb/cockroach/pkg/sql/exec/types"
-)
-
-func (m *memColumn) Append(
-	vec ColVec, colType types.T, toLength uint64, fromLength uint16,
-) {
-	switch colType {
-		{{range .}}
-		case types.{{.ExecType}}:
-			m.col = append(m.{{.ExecType}}()[:toLength], vec.{{.ExecType}}()[:fromLength]...)
-		{{end}}
-		default:
-			panic(fmt.Sprintf("unhandled type %d", colType))
-	}
-}
-
-func (m *memColumn) AppendWithSel(
-	vec ColVec, sel []uint16, batchSize uint16, colType types.T, toLength uint64,
-) {
-	switch colType {
-	{{range .}}
-	case types.{{.ExecType}}:
-		toCol := append(m.{{.ExecType}}()[:toLength], make([]{{.GoType}}, batchSize)...)
-		fromCol := vec.{{.ExecType}}()
-
-		for i := uint16(0); i < batchSize; i++ {
-      toCol[uint64(i) + toLength] = fromCol[sel[i]]
-		}
-
-		m.col = toCol
-	{{end}}
-	default:
-		panic(fmt.Sprintf("unhandled type %d", colType))
-	}
-}
-
-func (m *memColumn) Copy(src ColVec, srcStartIdx, srcEndIdx uint64, typ types.T) {
-	switch typ {
-	{{range .}}
-	case types.{{.ExecType}}:
-		copy(m.{{.ExecType}}(), src.{{.ExecType}}()[srcStartIdx:srcEndIdx])
-	{{end}}
-	default:
-		panic(fmt.Sprintf("unhandled type %d", typ))
-	}
-}
-
-func (m *memColumn) CopyWithSelInt64(
-	vec ColVec, sel []uint64, nSel uint16, colType types.T,
-) {
-	// todo (changangela): handle the case when nSel > ColBatchSize
-	switch colType {
-	{{range .}}
-	case types.{{.ExecType}}:
-		toCol := m.{{.ExecType}}()
-		fromCol := vec.{{.ExecType}}()
-		for i := uint16(0); i < nSel; i++ {
-			toCol[i] = fromCol[sel[i]]
-		}
-	{{end}}
-	default:
-		panic(fmt.Sprintf("unhandled type %d", colType))
-	}
-}
-
-func (m *memColumn) CopyWithSelInt16(vec ColVec, sel []uint16, nSel uint16, colType types.T) {
-	switch colType {
-	{{range .}}
-	case types.{{.ExecType}}:
-		toCol := m.{{.ExecType}}()
-		fromCol := vec.{{.ExecType}}()
-		for i := uint16(0); i < nSel; i++ {
-			toCol[i] = fromCol[sel[i]]
-		}
-	{{end}}
-	default:
-		panic(fmt.Sprintf("unhandled type %d", colType))
-	}
-}
-`
-
-type colVecGen struct {
-	ExecType string
-	GoType   string
-}
-
-func genColVec(wr io.Writer) error {
-	// Build list of all exec types.
-	var gens []colVecGen
-	for _, t := range types.AllTypes {
-		gen := colVecGen{
-			ExecType: t.String(),
-			GoType:   t.GoTypeName(),
-		}
-		gens = append(gens, gen)
-	}
-
-	tmpl, err := template.New("colVecTemplate").Parse(colVecTemplate)
+func genColvec(wr io.Writer) error {
+	d, err := ioutil.ReadFile("pkg/sql/exec/colvec_tmpl.go")
 	if err != nil {
 		return err
 	}
 
-	return tmpl.Execute(wr, gens)
-}
+	s := string(d)
 
+	// Replace the template variables.
+	s = strings.Replace(s, "_GOTYPE", "{{.LTyp.GoTypeName}}", -1)
+	s = strings.Replace(s, "_TYPES_T", "types.{{.LTyp}}", -1)
+	s = strings.Replace(s, "_TemplateType", "{{.LTyp}}", -1)
+
+	// Now, generate the op, from the template.
+	tmpl, err := template.New("colvec_op").Parse(s)
+	if err != nil {
+		return err
+	}
+
+	return tmpl.Execute(wr, comparisonOpToOverloads[tree.NE])
+}
 func init() {
-	registerGenerator(genColVec, "colvec.eg.go")
+	registerGenerator(genColvec, "colvec.eg.go")
 }
